backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.1
pydantic-settings==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.9
itsdangerous==2.2.0
aiofiles==23.2.1
orjson==3.10.7
